Refactors ssml computations to speech util.
